[Refactor/#138] MyChallengeDetailUiState를 model 패키지로 이동
